fix: example contracts fee: `stake` → `untrn`
