+++ contracts/ibc_transfer/src/lib.rs
@@ -1,4 +1,5 @@
 #![warn(clippy::unwrap_used, clippy::expect_used)]
 
 pub mod contract;
+pub mod msg;
 pub mod state;
