+++ scripts/test_interchain_txs.sh
@@ -1,53 +1,110 @@
-BIN=neutrond
-GAIA_BIN=gaiad
+#!/usr/bin/env bash
 
-CONTRACT=../artifacts/neutron_interchain_txs.wasm
+# http://redsymbol.net/articles/unofficial-bash-strict-mode/
+set -euo pipefail
+IFS=$'\n\t'
 
-CHAIN_ID_1=test-1
-CHAIN_ID_2=test-2
+BIN="neutrond"
+GAIA_BIN="gaiad"
+CONTRACT_PATH="../artifacts/neutron_interchain_txs.wasm"
+CHAIN_ID_1="test-1"
+CHAIN_ID_2="test-2"
+NEUTRON_DIR="${NEUTRON_DIR:-../../neutron}"
+HOME_1="${NEUTRON_DIR}/data/test-1/"
+HOME_2="${NEUTRON_DIR}/data/test-2/"
+ADDRESS_1="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
+ADDRESS_2="cosmos10h9stc5v6ntgeygf5xf945njqq5h32r53uquvw"
+ADMIN="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
+VALIDATOR="cosmosvaloper18hl5c9xn5dze2g50uaw0l2mr02ew57zk0auktn"
+NEUTRON_NODE="tcp://127.0.0.1:26657"
+GAIA_NODE="tcp://127.0.0.1:16657"
 
-NEUTRON_DIR=${NEUTRON_DIR:-../../neutron}
-HOME_1=${NEUTRON_DIR}/data/test-1/
-HOME_2=${NEUTRON_DIR}/data/test-2/
+code_id="$("$BIN" tx wasm store "$CONTRACT_PATH"                \
+    --from "$ADDRESS_1" --gas 50000000 --chain-id "$CHAIN_ID_1" \
+    --broadcast-mode=block --gas-prices 0.0025untrn -y          \
+    --output json --keyring-backend=test --home "$HOME_1"       \
+    --node "$NEUTRON_NODE"                                      \
+    | jq -r '.logs[0].events[] | select(.type == "store_code").attributes[] | select(.key == "code_id").value')"
+echo "Code ID: $code_id"
 
-ADDRESS_1=neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2
-ADDRESS_2=cosmos10h9stc5v6ntgeygf5xf945njqq5h32r53uquvw
-ADMIN=neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2
+contract_address=$("$BIN" tx wasm instantiate "$code_id" '{}'        \
+    --from "$ADDRESS_1" --admin "$ADMIN" -y --chain-id "$CHAIN_ID_1" \
+    --output json --broadcast-mode=block --label "init"              \
+    --keyring-backend=test --gas-prices 0.0025untrn --gas auto       \
+    --gas-adjustment 1.4 --home "$HOME_1"                            \
+    --node "$NEUTRON_NODE" 2>/dev/null                               \
+    | jq -r '.logs[0].events[] | select(.type == "instantiate").attributes[] | select(.key == "_contract_address").value')
+echo "Contract address: $contract_address"
 
-VAL2=cosmos1qnk2n4nlkpw9xfqntladh74w6ujtulwn7j8za9
+tx_result="$("$BIN" tx bank send demowallet1 "$contract_address" 100000untrn \
+    --chain-id "$CHAIN_ID_1" --home "$HOME_1" --node "$NEUTRON_NODE"         \
+    --keyring-backend=test -y --gas-prices 0.0025untrn                       \
+    --broadcast-mode=block --output json)"
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to send money to contract: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
+echo "Sent money to contract to pay fees"
 
-# Upload the txs contract
-RES=$(${BIN} tx wasm store ${CONTRACT} --from ${ADDRESS_1} --gas 50000000  --chain-id ${CHAIN_ID_1} --broadcast-mode=block --gas-prices 0.0025stake  -y --output json  --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-CONTRACT_CODE_ID=$(echo $RES | jq -r '.logs[0].events[1].attributes[0].value')
-echo $RES
-echo $CONTRACT_CODE_ID
+msg='{"register":{
+  "connection_id": "connection-0",
+  "interchain_account_id": "test"
+}}'
+tx_result="$("$BIN" tx wasm execute "$contract_address" "$msg"    \
+    --from "$ADDRESS_1" -y --chain-id "$CHAIN_ID_1" --output json \
+    --broadcast-mode=block --gas-prices 0.0025untrn --gas 1000000 \
+    --keyring-backend=test --home "$HOME_1" --node "$NEUTRON_NODE")"
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to register interchain account: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
+echo "Waiting 60 seconds for interchain account (sometimes it takes a lot of time)…"
+# shellcheck disable=SC2034
+for i in $(seq 60); do
+  sleep 1
+  echo -n .
+done
+echo " done"
 
-# Instantiate the contract
-INIT_CONTRACT='{}'
-echo "Instantiate"
-RES=$(${BIN} tx wasm instantiate $CONTRACT_CODE_ID "$INIT_CONTRACT" --from ${ADDRESS_1} --admin ${ADMIN} -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --label "init"  --keyring-backend test --gas-prices 0.0025stake --gas auto --gas-adjustment 1.4 --home ${HOME_1} --node tcp://127.0.0.1:16657)
-CONTRACT_ADDRESS=$(echo $RES | jq -r '.logs[0].events[0].attributes[0].value')
-echo $CONTRACT_ADDRESS
+# FIXME: why do we even perform a query like this, can't we do `neutrond query wasm smart blah-blah`?
+query='{"interchain_account_address_from_contract":{"interchain_account_id":"test"}}'
+query_b64_urlenc="$(echo -n "$query" | base64 | tr -d '\n' | jq -sRr '@uri')"
+url="http://127.0.0.1:1317/wasm/contract/$contract_address/smart/$query_b64_urlenc?encoding=base64"
+ica_address=$(curl -s "$url" | jq -r '.result.smart' | base64 -d | jq -r '.[0]')
+echo "ICA address: $ica_address"
 
-${BIN} tx bank send demowallet1 ${CONTRACT_ADDRESS} 100000stake --chain-id ${CHAIN_ID_1} --home ${HOME_1} --node tcp://localhost:16657 --keyring-backend test -y --gas-prices 0.0025stake --broadcast-mode=block
+tx_result=$("$GAIA_BIN" tx bank send "$ADDRESS_2" "$ica_address" 50000uatom  \
+    --chain-id "$CHAIN_ID_2" --broadcast-mode=block --gas-prices 0.0025uatom \
+    -y --output json --keyring-backend=test --home "$HOME_2" --node "$GAIA_NODE")
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to send money to ICA: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
+echo "Sent money to ICA"
 
-#Register interchain account
-RES=$(${BIN} tx wasm execute $CONTRACT_ADDRESS "{\"register\": {\"connection_id\": \"connection-0\", \"interchain_account_id\": \"test\"}}" --from ${ADDRESS_1}  -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --gas-prices 0.0025stake --gas 1000000 --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-echo $RES
-sleep 20
+msg="$(printf '{"delegate":{
+  "interchain_account_id": "test",
+  "validator": "%s",
+  "amount": "2000",
+  "denom": "uatom"
+}}' "$VALIDATOR")"
+tx_result="$("$BIN" tx wasm execute "$contract_address" "$msg"    \
+    --from "$ADDRESS_1" -y --chain-id "$CHAIN_ID_1" --output json \
+    --broadcast-mode=block --gas-prices 0.0025untrn --gas 1000000 \
+    --keyring-backend=test --home "$HOME_1" --node "$NEUTRON_NODE")"
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to execute contract: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
 
-RES=$(curl http://127.0.0.1:1316/wasm/contract/$CONTRACT_ADDRESS/smart/eyJpbnRlcmNoYWluX2FjY291bnRfYWRkcmVzc19mcm9tX2NvbnRyYWN0Ijp7ImludGVyY2hhaW5fYWNjb3VudF9pZCI6InRlc3QifX0\=?encoding\=base64 | jq -r ".result.smart")
-echo $RES
-ICA_ADDRESS=$(echo $RES | base64 --decode | jq -r ".[0]")
-echo $ICA_ADDRESS
+echo "Waiting 20 seconds for interchain transaction to complete…"
+# shellcheck disable=SC2034
+for i in $(seq 20); do
+  sleep 1
+  echo -n .
+done
+echo " done"
 
-#Send some money to ICA
-RES=$(${GAIA_BIN} tx bank send ${ADDRESS_2} ${ICA_ADDRESS} 10000stake --chain-id ${CHAIN_ID_2}  --broadcast-mode=block --gas-prices 0.0025stake -y --output json --keyring-backend test --home ${HOME_2} --node tcp://127.0.0.1:26657)
-echo $RES
-
-#Delegate
-RES=$(${BIN} tx wasm execute $CONTRACT_ADDRESS "{\"delegate\": {\"interchain_account_id\": \"test\", \"validator\": \"${VAL2}\", \"amount\":\"5000\",\"denom\":\"stake\"}}" --from ${ADDRESS_1}  -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --gas-prices 0.0025stake --gas 1000000 --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-echo $RES
-
-sleep 7;
-curl http://127.0.0.1:1317/staking/delegators/$ICA_ADDRESS/delegations
+echo
+echo "This delegation is performed using interchain tx module"
+curl -s "http://127.0.0.1:1316/staking/delegators/$ica_address/delegations" | jq '.result'
