+++ scripts/test_kv_query.sh
@@ -1,46 +1,68 @@
-BIN=neutrond
+#!/usr/bin/env bash
 
-CONTRACT=../artifacts/neutron_interchain_queries.wasm
+# http://redsymbol.net/articles/unofficial-bash-strict-mode/
+set -euo pipefail
+IFS=$'\n\t'
 
-CHAIN_ID_1=test-1
-CHAIN_ID_2=test-2
+BIN="neutrond"
+CONTRACT_PATH="../artifacts/neutron_interchain_queries.wasm"
+CHAIN_ID_1="test-1"
+NEUTRON_DIR="${NEUTRON_DIR:-../../neutron}"
+HOME_1="${NEUTRON_DIR}/data/test-1/"
+ADDRESS_1="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
+ADMIN="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
+NODE="tcp://127.0.0.1:26657"
 
-NEUTRON_DIR=${NEUTRON_DIR:-../../neutron}
-HOME_1=${NEUTRON_DIR}/data/test-1/
-HOME_2=${NEUTRON_DIR}/data/test-2/
+code_id="$("$BIN" tx wasm store "$CONTRACT_PATH"                \
+    --from "$ADDRESS_1" --gas 50000000 --chain-id "$CHAIN_ID_1" \
+    --broadcast-mode=block --gas-prices 0.0025untrn -y          \
+    --output json --keyring-backend=test --home "$HOME_1"       \
+    --node "$NODE"                                              \
+    | jq -r '.logs[0].events[] | select(.type == "store_code").attributes[] | select(.key == "code_id").value')"
+echo "Code ID: $code_id"
 
-ADDRESS_1=neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2
-ADDRESS_2=cosmos10h9stc5v6ntgeygf5xf945njqq5h32r53uquvw
-ADMIN=neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2
+contract_address="$("$BIN" tx wasm instantiate "$code_id" '{}'         \
+    --from "$ADDRESS_1" --admin "$ADMIN" -y --chain-id "$CHAIN_ID_1"   \
+    --output json --broadcast-mode=block --label "init" --node "$NODE" \
+    --keyring-backend=test --gas-prices 0.0025untrn --home "$HOME_1"   \
+    | jq -r '.logs[0].events[] | select(.type == "instantiate").attributes[] | select(.key == "_contract_address").value')"
+echo "Contract address: $contract_address"
 
-VAL2=cosmos1qnk2n4nlkpw9xfqntladh74w6ujtulwn7j8za9
+tx_result="$("$BIN" tx bank send "$ADDRESS_1" "$contract_address" 10000000untrn \
+    -y --chain-id "$CHAIN_ID_1" --output json --broadcast-mode=block            \
+    --gas-prices 0.0025untrn --gas 300000 --keyring-backend=test                \
+    --home "$HOME_1" --node "$NODE")"
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to send money to contract: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
+echo "Sent money to contract to pay for deposit"
 
-# Upload the queries contract
-echo "Upload the queries contract"
-RES=$(${BIN} tx wasm store ${CONTRACT} --from ${ADDRESS_1} --gas 50000000  --chain-id ${CHAIN_ID_1} --broadcast-mode=block --gas-prices 0.0025stake  -y --output json  --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-QUERIES_CONTRACT_CODE_ID=$(echo $RES | jq -r '.logs[0].events[1].attributes[0].value')
-echo $RES
-echo $QUERIES_CONTRACT_CODE_ID
+msg='{"register_bank_total_supply_query":{
+  "connection_id": "connection-0",
+  "denoms": ["uatom"],
+  "update_period": 5
+}}'
+tx_result="$("$BIN" tx wasm execute "$contract_address" "$msg"    \
+    --from "$ADDRESS_1" -y --chain-id "$CHAIN_ID_1" --output json \
+    --broadcast-mode=block --gas-prices 0.0025untrn --gas 1000000 \
+    --keyring-backend=test --home "$HOME_1" --node "$NODE")"
+code="$(echo "$tx_result" | jq '.code')"
+if [[ "$code" -ne 0 ]]; then
+  echo "Failed to register ICQ: $(echo "$tx_result" | jq '.raw_log')" && exit 1
+fi
+query_id="$(echo "$tx_result" | jq -r '.logs[0].events[] | select(.type == "neutron").attributes[] | select(.key == "query_id").value')"
+echo "Registered total supply ICQ with query ID: $query_id"
 
-# Instantiate the queries contract
-echo "Instantiate the queries contract"
-INIT_QUERIES_CONTRACT='{}'
+echo "Waiting 10 seconds for ICQ result to arrive…"
+# shellcheck disable=SC2034
+for i in $(seq 10); do
+  sleep 1
+  echo -n .
+done
+echo " done"
 
-RES=$(${BIN} tx wasm instantiate $QUERIES_CONTRACT_CODE_ID "$INIT_QUERIES_CONTRACT" --from ${ADDRESS_1} --admin ${ADMIN} -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --label "init"  --keyring-backend test --gas-prices 0.0025stake --home ${HOME_1} --node tcp://127.0.0.1:16657)
-echo $RES
-QUERIES_CONTRACT_ADDRESS=$(echo $RES | jq -r '.logs[0].events[0].attributes[0].value')
-echo $QUERIES_CONTRACT_ADDRESS
-
-# Send coins from USERNAME_1 to QUERIES_CONTRACT_ADDRESS to perform register_interchain_query message
-echo "Send coins from ${ADDRESS_1} to ${QUERIES_CONTRACT_ADDRESS} to perform register_interchain_query message"
-echo $(${BIN} tx bank send ${ADDRESS_1} ${QUERIES_CONTRACT_ADDRESS} 10000000stake -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --gas-prices 0.0025stake --gas 300000 --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-
-# Register a query for a total supply of stake token
-echo "Register a query for total supply"
-RES=$(${BIN} tx wasm execute $QUERIES_CONTRACT_ADDRESS "{\"register_bank_total_supply_query\": {\"connection_id\": \"connection-0\", \"denoms\": [\"stake\"], \"update_period\": 5}}" --from ${ADDRESS_1}  -y --chain-id ${CHAIN_ID_1} --output json --broadcast-mode=block --gas-prices 0.0025stake --gas 1000000 --keyring-backend test --home ${HOME_1} --node tcp://127.0.0.1:16657)
-echo $RES | jq
-
-sleep 5
-
-echo "KV query response:"
-${BIN} query wasm contract-state smart ${QUERIES_CONTRACT_ADDRESS} '{"bank_total_supply": {"query_id": 1}}' --node tcp://127.0.0.1:16657
+echo
+echo "KV query total supply response:"
+query="$(printf '{"bank_total_supply": {"query_id": %s}}' "$query_id")"
+"$BIN" query wasm contract-state smart "$contract_address" "$query" --node "$NODE" --output json | jq
